+++ CHANGELOG.md
@@ -1,3 +1,21 @@
 ## 0.0.1
 
-* Plugin initialiazation
+* Plugin initialization
+  
+  Supported features:
+    - notification receiving with DLRs
+    - seen reports on notification click event
+    - notification settings
+    - privacy settings
+    - Installation (device) data default sync
+    - rich notifications
+    - interactive notifications
+    - in-app notifications
+
+## 0.0.2
+
+* Fixed major bug at Android build
+
+## 0.1.0
+
+* Support for User and Installation management (personalization).